Tidy up imports
It is idiomatic in Rust to import the containing module of a function
and then use it by first referring to the module, like so:
`module::function`. Types, on the other hand, are typically imported
directly. Exceptions to this convention include the case where importing
a type will clobber an existing type (such as with Rust's `Err` and
`nom::Err`). See the relevant section in the Rust book for more
information: https://doc.rust-lang.org/book/ch07-04-bringing-paths-into-scope-with-the-use-keyword.html#creating-idiomatic-use-paths